Adjustement to pAus4777.py and result.py to adapte the graph
the changes to result.py consist of being able to send a minimal value for x axis. if not sent througth get_rslt_param_plot, will start at 0
and the x_axis consist of the voltage values instead of Time. result.py is not currently able to support two variables.